% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_advanced_search.R
\name{core_advanced_search}
\alias{core_advanced_search}
\alias{core_advanced_search_}
\title{Advanced Search CORE}
\usage{
core_advanced_search(query, page = 1, limit = 10, key = NULL,
  parse = TRUE, ...)

core_advanced_search_(query, page = 1, limit = 10, key = NULL, ...)
}
\arguments{
\item{query}{data.frame, required}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10),
optional}

\item{key}{A IUCN API token}

\item{parse}{(logical) Whether to parse to list (`FALSE`) or
data.frame (`TRUE`). Default: `TRUE`}

\item{...}{Curl options passed to [crul::HttpClient()]}
}
\description{
Advanced Search CORE
}
\details{
\code{core_advanced_search} does the HTTP request and parses, while
\code{core_advanced_search_} just does the HTTP request, gives back JSON as a character
string
}
\references{
\url{https://core.ac.uk/docs/#!/all/advanced_search}
}
