% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_repos_search.R, R/core_search.R
\name{core_repos_search_}
\alias{core_repos_search_}
\alias{core_search}
\alias{core_search_}
\title{Search CORE}
\usage{
core_repos_search_(query, page = 1, limit = 10, key = NULL, ...)

core_search(query, page = 1, limit = 10, key = NULL, parse = TRUE, ...)

core_search_(query, page = 1, limit = 10, key = NULL, ...)
}
\arguments{
\item{query}{(character) query string, required}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}

\item{key}{A IUCN API token}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}
}
\description{
Search CORE
}
\details{
\code{core_search} does the HTTP request and parses, while
\code{core_search_} just does the HTTP request, gives back JSON as a character
string
}
\examples{
\dontrun{
core_search(query = 'ecology')
core_search(query = 'ecology', parse = FALSE)
core_search(query = 'ecology', limit = 2)

core_search_(query = 'ecology')
library("jsonlite")
jsonlite::fromJSON(core_search_(query = 'ecology'))

# core_search_(query = 'ecology', limit = 3)
}
}
\references{
\url{https://core.ac.uk/docs/#!/all/search}
}

