% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cored-package.R
\docType{package}
\name{cored-package}
\alias{cored-package}
\alias{cored}
\title{cored - CORE R client}
\description{
cored - CORE R client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
