% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_articles_pdf.R
\name{core_articles_pdf}
\alias{core_articles_pdf}
\alias{core_articles_pdf_}
\title{Download article pdf}
\usage{
core_articles_pdf(id, key = NULL, overwrite = TRUE, parse = TRUE, ...)

core_articles_pdf_(id, key = NULL, overwrite = TRUE, ...)
}
\arguments{
\item{id}{(integer) CORE ID of the article that needs to be fetched.
Required}

\item{key}{A IUCN API token}

\item{overwrite}{(logical) overwrite file or not. Default: \code{TRUE}}

\item{parse}{(logical) Whether to parse to list (`FALSE`) or
data.frame (`TRUE`). Default: `TRUE`}

\item{...}{Curl options passed to [crul::HttpClient()]}
}
\description{
Download article pdf
}
\details{
\code{core_articles_pdf} does the HTTP request and parses
PDF to text, while \code{core_articles_pdf_} just does the HTTP request
and gives back the path to the file

We now do a request to \code{\link[=core_articles]{core_articles()}} first to find out if there is a
PDF available via existence of the field \code{fulltextIdentifier}. If not,
we stop with error message; if the field exists, we proceed to
attempting to fetch the PDF.

These functions take one article ID at a time. Use lapply/loops/etc for
many ids
}
\examples{
\dontrun{
# just http request, get file path back
core_articles_pdf_(11549557)

# get paper and parse to text
core_articles_pdf(11549557)

ids <- c(11549557, 385071)
res <- lapply(ids, core_articles_pdf)
vapply(res, "[[", "", 1)
}
}
\references{
\url{https://core.ac.uk/docs/#!/articles/getArticlePdfByCoreId}
}
